✨ *feat(bot): implementar nova estrutura de tarefas para o bot*
Esta atualização introduz uma nova estrutura de tarefas para o bot, melhorando a organização e a eficiência do código.

Files Changed:
- *celery_app/__init__.py*: Adicionada a instância do Celery e reorganização da função make_celery.
- *celery_app/__main__.py*: Alterado o nome da importação para manter a consistência.
- *celery_app/custom/_celery.py*: Removido método desnecessário e ajustada a estrutura da classe AsyncCelery.
- *celery_app/main.py*: Importada a instância do Celery para uso no módulo principal.
- *celery_app/tasks/__init__.py*: Registrada a nova tarefa BotTask e removido o antigo pje.
- *celery_app/tasks/bot/_main.py*: Criada nova tarefa para inicialização e execução de instâncias de bot.
- *celery_app/tasks/bot/message.py*: Implementada nova tarefa para comunicação via Socket.IO.
- *crawjud/_wrapper.py*: Adicionada função para envolver a inicialização de classes de bot.
- *crawjud/bots/pje/movimentacao.py*: Atualizada a classe Movimentacao para herdar de ClassBot e removido método desnecessário.
- *crawjud/core/__init__.py*: Reestruturada a classe CrawJUD, removendo herança desnecessária e ajustando a inicialização.
- *crawjud/core/master.py*: Removido código obsoleto relacionado à classe Controller.
